Removes AASEndpointUtil class
